+++ src/Aydsko.iRacingTelemetry/Aydsko.iRacingTelemetry.Generators.Attributes/Aydsko.iRacingTelemetry.Generators.Attributes.csproj
@@ -1,0 +1,7 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+</Project>
