+++ src/Aydsko.iRacingTelemetry/Aydsko.iRacingTelemetry.Generators/Aydsko.iRacingTelemetry.Generators.csproj
