+++ src/Aydsko.iRacingTelemetry/Aydsko.iRacingTelemetry.Generators/Mapper/AydskoMapperContext.cs
