+++ src/Aydsko.iRacingTelemetry/Aydsko.iRacingTelemetry.UnitTests/Aydsko.iRacingTelemetry.UnitTests.csproj
@@ -10,11 +10,6 @@
     <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
   </PropertyGroup>
 
-  <PropertyGroup>
-    <EmitCompilerGeneratedFiles>true</EmitCompilerGeneratedFiles>
-    <CompilerGeneratedFilesOutputPath>Generated</CompilerGeneratedFilesOutputPath>
-  </PropertyGroup>
-
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.7.2" />
     <PackageReference Include="NUnit" Version="3.13.3" />
@@ -31,10 +26,12 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Aydsko.iRacingTelemetry\Aydsko.iRacingTelemetry.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Aysdko.iRacingTelemetry.Generators\Aydsko.iRacingTelemetry.Generators.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="true" />
+    <ProjectReference Include="..\Aydsko.iRacingTelemetry.Generators\Aydsko.iRacingTelemetry.Generators.csproj"
+                      OutputItemType="Analyzer"
+                      ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\Aydsko.iRacingTelemetry.Generators.Attributes\Aydsko.iRacingTelemetry.Generators.Attributes.csproj"
+                      OutputItemType="Analyzer"
+                      ReferenceOutputAssembly="true" />
   </ItemGroup>
 
   <ItemGroup>
