+++ src/Aydsko.iRacingTelemetry/Aydsko.iRacingTelemetry.sln
@@ -13,7 +13,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Aydsko.iRacingTelemetry.UnitTests", "Aydsko.iRacingTelemetry.UnitTests\Aydsko.iRacingTelemetry.UnitTests.csproj", "{DB276308-5C05-4EEC-9914-B11AD3F37DEA}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Aydsko.iRacingTelemetry.Generators", "Aysdko.iRacingTelemetry.Generators\Aydsko.iRacingTelemetry.Generators.csproj", "{1BDF487C-4E58-424A-8BBE-0C2A74BE0AA6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Aydsko.iRacingTelemetry.Generators.Attributes", "Aydsko.iRacingTelemetry.Generators.Attributes\Aydsko.iRacingTelemetry.Generators.Attributes.csproj", "{3D291C41-B5E5-4AC4-B5E6-5D270D82C026}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Aydsko.iRacingTelemetry.Generators", "Aydsko.iRacingTelemetry.Generators\Aydsko.iRacingTelemetry.Generators.csproj", "{06FE08A2-9B84-4569-BADD-0D0FDB561BD0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,10 +31,14 @@ Global
 		{DB276308-5C05-4EEC-9914-B11AD3F37DEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DB276308-5C05-4EEC-9914-B11AD3F37DEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DB276308-5C05-4EEC-9914-B11AD3F37DEA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1BDF487C-4E58-424A-8BBE-0C2A74BE0AA6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1BDF487C-4E58-424A-8BBE-0C2A74BE0AA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1BDF487C-4E58-424A-8BBE-0C2A74BE0AA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1BDF487C-4E58-424A-8BBE-0C2A74BE0AA6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3D291C41-B5E5-4AC4-B5E6-5D270D82C026}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3D291C41-B5E5-4AC4-B5E6-5D270D82C026}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3D291C41-B5E5-4AC4-B5E6-5D270D82C026}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3D291C41-B5E5-4AC4-B5E6-5D270D82C026}.Release|Any CPU.Build.0 = Release|Any CPU
+		{06FE08A2-9B84-4569-BADD-0D0FDB561BD0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{06FE08A2-9B84-4569-BADD-0D0FDB561BD0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{06FE08A2-9B84-4569-BADD-0D0FDB561BD0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{06FE08A2-9B84-4569-BADD-0D0FDB561BD0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
