--- src/Aydsko.iRacingTelemetry/Aysdko.iRacingTelemetry.Generators/Mapper/AydskoMapAttribute.cs
@@ -1,12 +1,0 @@
-﻿namespace Aydsko.iRacingTelemetry.Generators.Mapper;
-
-[AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
-public sealed class AydskoMapAttribute : Attribute
-{
-    public string? IrsdkVariableName { get; }
-
-    public AydskoMapAttribute(string? irsdkVariableName = null)
-    {
-        IrsdkVariableName = irsdkVariableName;
-    }
-}
