--- src/Aydsko.iRacingTelemetry/Aysdko.iRacingTelemetry.Generators/Mapper/IAydskoMapper.cs
@@ -1,6 +1,0 @@
-﻿namespace Aydsko.iRacingTelemetry.Generators.Mapper;
-
-public interface IAydskoMapper<T> where T : class
-{
-    T MapFromBuffer(Span<byte> buffer);
-}
